Tempo: Move interfaces & `TempoQueryType` from `datasource.ts` to `types.ts` (#54621)
