+++ public/app/plugins/datasource/tempo/QueryEditor/ServiceGraphSection.tsx
@@ -9,7 +9,7 @@ import { Alert, InlineField, InlineFieldRow, useStyles2 } from '@grafana/ui';
 import { AdHocFilter } from '../../../../features/variables/adhoc/picker/AdHocFilter';
 import { AdHocVariableFilter } from '../../../../features/variables/types';
 import { PrometheusDatasource } from '../../prometheus/datasource';
-import { TempoQuery } from '../datasource';
+import { TempoQuery } from '../types';
 
 import { getDS } from './utils';
