+++ public/app/plugins/datasource/tempo/configuration/LokiSearchSettings.tsx
@@ -5,7 +5,7 @@ import { DataSourcePluginOptionsEditorProps, GrafanaTheme, updateDatasourcePlugi
 import { DataSourcePicker } from '@grafana/runtime';
 import { Button, InlineField, InlineFieldRow, useStyles } from '@grafana/ui';
 
-import { TempoJsonData } from '../datasource';
+import { TempoJsonData } from '../types';
 
 interface Props extends DataSourcePluginOptionsEditorProps<TempoJsonData> {}
